+++ lib/domain/settings/reminder_times_page.dart
@@ -1,6 +1,6 @@
 import 'package:pilll/components/molecules/indicator.dart';
 import 'package:pilll/entity/setting.dart';
-import 'package:pilll/store/setting.dart';
+import 'package:pilll/domain/settings/setting_page_store.dart';
 import 'package:pilll/components/atoms/color.dart';
 import 'package:pilll/components/atoms/font.dart';
 import 'package:pilll/components/atoms/text_color.dart';
