+++ lib/domain/settings/settings_page.dart
@@ -14,7 +14,7 @@ import 'package:pilll/domain/settings/reminder_times_page.dart';
 import 'package:pilll/error/error_alert.dart';
 import 'package:pilll/inquiry/inquiry.dart';
 import 'package:pilll/service/pill_sheet.dart';
-import 'package:pilll/store/setting.dart';
+import 'package:pilll/domain/settings/setting_page_store.dart';
 import 'package:pilll/components/atoms/color.dart';
 import 'package:pilll/components/atoms/font.dart';
 import 'package:pilll/components/atoms/text_color.dart';
