+++ test/domain/settings/reminder_times_widget_test.dart
@@ -1,7 +1,7 @@
 import 'package:pilll/entity/pill_sheet_type.dart';
 import 'package:pilll/entity/setting.dart';
 import 'package:pilll/domain/settings/reminder_times_page.dart';
-import 'package:pilll/store/setting.dart';
+import 'package:pilll/domain/settings/setting_page_store.dart';
 import 'package:pilll/util/environment.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
